+++ packages/ui/src/components/ui/type-badges/index.tsx
@@ -2,16 +2,17 @@ import clsx from 'clsx';
 import { TypeNames } from 'contract';
 import { Badge } from '../badge';
 import { Tiny } from '../typography';
-import { commonClasseses, typeClassNames } from './helpers';
+import { commonClasseses, tinyClasseses, typeClassNames } from './helpers';
 
 interface TypeBadgeProps {
   type: TypeNames;
+  tiny?: boolean;
 }
 
-export function TypeBadge({ type = 'empty' }: TypeBadgeProps) {
+export function TypeBadge({ type = 'empty', tiny = false }: TypeBadgeProps) {
   return (
-    <Badge variant='default' className={clsx(commonClasseses, typeClassNames[type])}>
-      <Tiny>{type}</Tiny>
+    <Badge variant='default' className={clsx(tiny ? tinyClasseses : commonClasseses, typeClassNames[type])}>
+      <Tiny>{tiny ? type[0] : type}</Tiny>
     </Badge>
   );
 }
